[update] messaging, [update] entities, [update] instalaciones

+++ src/entities/Posiciones.ts
@@ -86,7 +86,7 @@ export class Posiciones {
   @Column('int', { name: 'Movimiento', nullable: true })
   movimiento: number | null;
 
-  @Column('int', { name: 'Combustible', nullable: true })
+  @Column('float', { name: 'Combustible', nullable: true })
   combustible: number | null;
 
   @Column('bigint', { name: 'IdFoto1', nullable: true })
